feat : Service 추상화 계층 삭제 (#3)
